Refactor all of the primitive components into components/ui folder
