+++ front-end/src/components/ui/Button/ButtonProps.tsx
@@ -1,5 +1,5 @@
 import { VariantProps, cva } from 'class-variance-authority';
-import { cn } from '../../utils/utils';
+import { cn } from '../../../utils/utils';
 
 const buttonVariants = cva(
   cn(
