+++ front-end/src/components/ui/Icon/Icon.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '../../utils/utils';
+import { cn } from '../../../utils/utils';
 
 type Props = {
   name: string;
