+++ front-end/src/components/ui/Panel/Panel.tsx
