+++ front-end/src/components/ui/Switch/Switch.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { cn } from '../../utils/utils';
+import { cn } from '../../../utils/utils';
 
 type ButtonProps = {
   onClick: () => void;
