+++ front-end/src/components/ui/modal/Modal.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import ReactDOM from 'react-dom';
 import { cn } from '../../../utils/utils';
-import FocusTrap from '../../FocusTrap/FocusTrap';
+import FocusTrap from '../FocusTrap/FocusTrap';
 
 interface Props {
   children: React.ReactNode;
